Refactor product and category GraphQL resolvers and schemas
- Updated product-related resolvers to handle creation, updating, and deletion of products with improved input handling.
- Introduced new GraphQL queries and mutations for categories and collections, including pagination support.
- Replaced scalar.Decimal with shopspring/decimal for better precision in price handling.
- Cleaned up schema definitions by consolidating product, category, and collection operations into their respective types.
- Removed redundant resolver implementations from schema.resolvers.go.
- Updated .env.example to change log format from JSON to text for better readability.
