+++ gitstore-api/.env.example
@@ -20,7 +20,7 @@ GITSTORE_API__GIT_PORT=5000                     # int     — git smart HTTP lis
 GITSTORE_API_PORT=4000                    # int      — HTTP API listen port
 GITSTORE_CACHE__TTL=300                   # int      — catalog cache TTL in seconds
 GITSTORE_LOG__LEVEL=info                  # string   — debug | info | warn | error
-GITSTORE_LOG__FORMAT=json                 # string  — log format (json or text)
+GITSTORE_LOG__FORMAT=text                 # string  — log format (json or text)
 GITSTORE_AUTH__JWT__DURATION=24h          # duration — JWT token lifetime
 GITSTORE_AUTH__JWT__ISSUER=gitstore       # string   — JWT issuer claim
