+++ gitstore-api/internal/graph/schema.resolvers.go
@@ -7,363 +7,11 @@ package graph
 
 import (
 	"context"
-	"fmt"
 	"time"
 
 	"github.com/gitstore-dev/gitstore/api/internal/graph/generated"
 	"github.com/gitstore-dev/gitstore/api/internal/graph/model"
 )
-
-// CreateProduct is the resolver for the createProduct field.
-func (r *mutationResolver) CreateProduct(ctx context.Context, input model.CreateProductInput) (*model.CreateProductPayload, error) {
-	categoryID, err := decodeNodeIDAs(nodeKindCategory, input.CategoryID)
-	if err != nil {
-		return nil, err
-	}
-	collectionIDs, err := decodeNodeIDsAs(nodeKindCollection, input.CollectionIds)
-	if err != nil {
-		return nil, err
-	}
-
-	// Build input map for service
-	priceFloat, _ := input.Price.Float64()
-	serviceInput := map[string]interface{}{
-		"title":      input.Title,
-		"sku":        input.Sku,
-		"price":      priceFloat,
-		"categoryId": categoryID,
-	}
-
-	if input.Currency != nil {
-		serviceInput["currency"] = *input.Currency
-	}
-	if input.Body != nil {
-		serviceInput["body"] = *input.Body
-	}
-	if input.InventoryStatus != nil {
-		serviceInput["inventoryStatus"] = string(*input.InventoryStatus)
-	}
-	if input.InventoryQuantity != nil {
-		serviceInput["inventoryQuantity"] = int(*input.InventoryQuantity)
-	}
-	if len(input.CollectionIds) > 0 {
-		serviceInput["collectionIds"] = collectionIDs
-	}
-	if len(input.Images) > 0 {
-		serviceInput["images"] = input.Images
-	}
-	if input.Metadata != nil {
-		serviceInput["metadata"] = input.Metadata
-	}
-
-	// Create product via service
-	product, err := r.service.CreateProduct(ctx, serviceInput)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create product: %w", err)
-	}
-
-	return &model.CreateProductPayload{
-		ClientMutationID: input.ClientMutationID,
-		Product:          DatastoreProductToGraphQL(product),
-	}, nil
-}
-
-// UpdateProduct is the resolver for the updateProduct field.
-func (r *mutationResolver) UpdateProduct(ctx context.Context, input model.UpdateProductInput) (*model.UpdateProductPayload, error) {
-	productID, err := decodeNodeIDAs(nodeKindProduct, input.ID)
-	if err != nil {
-		return nil, err
-	}
-	categoryID, err := decodeOptionalNodeIDAs(nodeKindCategory, input.CategoryID)
-	if err != nil {
-		return nil, err
-	}
-	collectionIDs, err := decodeNodeIDsAs(nodeKindCollection, input.CollectionIds)
-	if err != nil {
-		return nil, err
-	}
-
-	// Build input map for service
-	serviceInput := map[string]interface{}{}
-
-	if input.Title != nil {
-		serviceInput["title"] = *input.Title
-	}
-	if input.Sku != nil {
-		serviceInput["sku"] = *input.Sku
-	}
-	if input.Price != nil {
-		priceFloat, _ := input.Price.Float64()
-		serviceInput["price"] = priceFloat
-	}
-	if input.Currency != nil {
-		serviceInput["currency"] = *input.Currency
-	}
-	if input.Body != nil {
-		serviceInput["body"] = *input.Body
-	}
-	if input.InventoryStatus != nil {
-		serviceInput["inventoryStatus"] = string(*input.InventoryStatus)
-	}
-	if input.InventoryQuantity != nil {
-		serviceInput["inventoryQuantity"] = int(*input.InventoryQuantity)
-	}
-	if input.CategoryID != nil {
-		serviceInput["categoryId"] = *categoryID
-	}
-	if len(input.CollectionIds) > 0 {
-		serviceInput["collectionIds"] = collectionIDs
-	}
-	if len(input.Images) > 0 {
-		serviceInput["images"] = input.Images
-	}
-	if input.Metadata != nil {
-		serviceInput["metadata"] = input.Metadata
-	}
-
-	// Update product via service
-	product, err := r.service.UpdateProduct(ctx, productID, serviceInput)
-	if err != nil {
-		return nil, fmt.Errorf("failed to update product: %w", err)
-	}
-
-	return &model.UpdateProductPayload{
-		ClientMutationID: input.ClientMutationID,
-		Product:          DatastoreProductToGraphQL(product),
-		Conflict:         nil, // TODO: Implement optimistic locking
-	}, nil
-}
-
-// DeleteProduct is the resolver for the deleteProduct field.
-func (r *mutationResolver) DeleteProduct(ctx context.Context, input model.DeleteProductInput) (*model.DeleteProductPayload, error) {
-	productID, err := decodeNodeIDAs(nodeKindProduct, input.ID)
-	if err != nil {
-		return nil, err
-	}
-
-	// Delete product via service
-	err = r.service.DeleteProduct(ctx, productID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to delete product: %w", err)
-	}
-
-	idCopy := mustEncodeNodeID(nodeKindProduct, productID)
-	return &model.DeleteProductPayload{
-		ClientMutationID: input.ClientMutationID,
-		DeletedProductID: &idCopy,
-	}, nil
-}
-
-// CreateCategory is the resolver for the createCategory field.
-func (r *mutationResolver) CreateCategory(ctx context.Context, input model.CreateCategoryInput) (*model.CreateCategoryPayload, error) {
-	parentID, err := decodeOptionalNodeIDAs(nodeKindCategory, input.ParentID)
-	if err != nil {
-		return nil, err
-	}
-
-	serviceInput := map[string]interface{}{
-		"name": input.Name,
-		"slug": input.Slug,
-	}
-	if input.Body != nil {
-		serviceInput["body"] = *input.Body
-	}
-	if parentID != nil {
-		serviceInput["parentId"] = *parentID
-	}
-	if input.DisplayOrder != nil {
-		serviceInput["displayOrder"] = *input.DisplayOrder
-	}
-
-	category, err := r.service.CreateCategory(ctx, serviceInput)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create category: %w", err)
-	}
-
-	return &model.CreateCategoryPayload{
-		ClientMutationID: input.ClientMutationID,
-		Category:         DatastoreCategoryToGraphQL(category),
-	}, nil
-}
-
-// UpdateCategory is the resolver for the updateCategory field.
-func (r *mutationResolver) UpdateCategory(ctx context.Context, input model.UpdateCategoryInput) (*model.UpdateCategoryPayload, error) {
-	categoryID, err := decodeNodeIDAs(nodeKindCategory, input.ID)
-	if err != nil {
-		return nil, err
-	}
-	if _, err := decodeOptionalNodeIDAs(nodeKindCategory, input.ParentID); err != nil {
-		return nil, err
-	}
-
-	// TODO: Implement proper git-backed category update
-	// For now, return mock response to unblock E2E tests
-	category := &model.Category{
-		ID:           mustEncodeNodeID(nodeKindCategory, categoryID),
-		Name:         stringOrDefault(input.Name, "Updated Category"),
-		Slug:         stringOrDefault(input.Slug, "updated-category"),
-		Body:         input.Body,
-		Parent:       nil, // TODO: lookup parent
-		Children:     []*model.Category{},
-		DisplayOrder: intOrDefault(input.DisplayOrder, 0),
-		Products:     &model.ProductConnection{Edges: []*model.ProductEdge{}, PageInfo: &model.PageInfo{}},
-		CreatedAt:    time.Now().Add(-24 * time.Hour),
-		UpdatedAt:    time.Now(),
-	}
-
-	return &model.UpdateCategoryPayload{
-		ClientMutationID: input.ClientMutationID,
-		Category:         category,
-		Conflict:         nil,
-	}, nil
-}
-
-// DeleteCategory is the resolver for the deleteCategory field.
-func (r *mutationResolver) DeleteCategory(ctx context.Context, input model.DeleteCategoryInput) (*model.DeleteCategoryPayload, error) {
-	categoryID, err := decodeNodeIDAs(nodeKindCategory, input.ID)
-	if err != nil {
-		return nil, err
-	}
-
-	// TODO: Implement proper git-backed category deletion
-	// For now, return mock response to unblock E2E tests
-	idCopy := mustEncodeNodeID(nodeKindCategory, categoryID)
-	return &model.DeleteCategoryPayload{
-		ClientMutationID:  input.ClientMutationID,
-		DeletedCategoryID: &idCopy,
-	}, nil
-}
-
-// ReorderCategories is the resolver for the reorderCategories field.
-func (r *mutationResolver) ReorderCategories(ctx context.Context, input model.ReorderCategoriesInput) (*model.ReorderCategoriesPayload, error) {
-	orderedIDs, err := decodeNodeIDsAs(nodeKindCategory, input.OrderedIds)
-	if err != nil {
-		return nil, err
-	}
-	if _, err := decodeOptionalNodeIDAs(nodeKindCategory, input.ParentID); err != nil {
-		return nil, err
-	}
-	if _, err := decodeOptionalNodeIDAs(nodeKindCategory, input.MovedCategoryID); err != nil {
-		return nil, err
-	}
-	if _, err := decodeOptionalNodeIDAs(nodeKindCategory, input.NewParentID); err != nil {
-		return nil, err
-	}
-
-	// TODO: Implement proper git-backed category reordering
-	// For now, return mock response with updated display orders
-	categories := make([]*model.Category, len(orderedIDs))
-	for i, id := range orderedIDs {
-		categories[i] = &model.Category{
-			ID:           mustEncodeNodeID(nodeKindCategory, id),
-			Name:         fmt.Sprintf("Category %d", i+1),
-			Slug:         fmt.Sprintf("category-%d", i+1),
-			DisplayOrder: int32(i),
-			Children:     []*model.Category{},
-			Products:     &model.ProductConnection{Edges: []*model.ProductEdge{}, PageInfo: &model.PageInfo{}},
-			CreatedAt:    time.Now(),
-			UpdatedAt:    time.Now(),
-		}
-	}
-
-	return &model.ReorderCategoriesPayload{
-		ClientMutationID: input.ClientMutationID,
-		Categories:       categories,
-	}, nil
-}
-
-// CreateCollection is the resolver for the createCollection field.
-func (r *mutationResolver) CreateCollection(ctx context.Context, input model.CreateCollectionInput) (*model.CreateCollectionPayload, error) {
-	if _, err := decodeNodeIDsAs(nodeKindProduct, input.ProductIds); err != nil {
-		return nil, err
-	}
-
-	serviceInput := map[string]interface{}{
-		"name": input.Name,
-		"slug": input.Slug,
-	}
-	if input.Body != nil {
-		serviceInput["body"] = *input.Body
-	}
-	if input.DisplayOrder != nil {
-		serviceInput["displayOrder"] = *input.DisplayOrder
-	}
-
-	collection, err := r.service.CreateCollection(ctx, serviceInput)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create collection: %w", err)
-	}
-
-	return &model.CreateCollectionPayload{
-		ClientMutationID: input.ClientMutationID,
-		Collection:       DatastoreCollectionToGraphQL(collection),
-	}, nil
-}
-
-// UpdateCollection is the resolver for the updateCollection field.
-func (r *mutationResolver) UpdateCollection(ctx context.Context, input model.UpdateCollectionInput) (*model.UpdateCollectionPayload, error) {
-	collectionID, err := decodeNodeIDAs(nodeKindCollection, input.ID)
-	if err != nil {
-		return nil, err
-	}
-	if _, err := decodeNodeIDsAs(nodeKindProduct, input.ProductIds); err != nil {
-		return nil, err
-	}
-
-	collection := &model.Collection{
-		ID:           mustEncodeNodeID(nodeKindCollection, collectionID),
-		Name:         stringOrDefault(input.Name, "Updated Collection"),
-		Slug:         stringOrDefault(input.Slug, "updated-collection"),
-		Body:         input.Body,
-		DisplayOrder: intOrDefault(input.DisplayOrder, 0),
-		Products:     &model.ProductConnection{Edges: []*model.ProductEdge{}, PageInfo: &model.PageInfo{}},
-		CreatedAt:    time.Now().Add(-24 * time.Hour),
-		UpdatedAt:    time.Now(),
-	}
-	return &model.UpdateCollectionPayload{
-		ClientMutationID: input.ClientMutationID,
-		Collection:       collection,
-		Conflict:         nil,
-	}, nil
-}
-
-// DeleteCollection is the resolver for the deleteCollection field.
-func (r *mutationResolver) DeleteCollection(ctx context.Context, input model.DeleteCollectionInput) (*model.DeleteCollectionPayload, error) {
-	collectionID, err := decodeNodeIDAs(nodeKindCollection, input.ID)
-	if err != nil {
-		return nil, err
-	}
-	idCopy := mustEncodeNodeID(nodeKindCollection, collectionID)
-	return &model.DeleteCollectionPayload{
-		ClientMutationID:    input.ClientMutationID,
-		DeletedCollectionID: &idCopy,
-	}, nil
-}
-
-// ReorderCollections is the resolver for the reorderCollections field.
-func (r *mutationResolver) ReorderCollections(ctx context.Context, input model.ReorderCollectionsInput) (*model.ReorderCollectionsPayload, error) {
-	orderedIDs, err := decodeNodeIDsAs(nodeKindCollection, input.OrderedIds)
-	if err != nil {
-		return nil, err
-	}
-
-	collections := make([]*model.Collection, len(orderedIDs))
-	for i, id := range orderedIDs {
-		collections[i] = &model.Collection{
-			ID:           mustEncodeNodeID(nodeKindCollection, id),
-			Name:         fmt.Sprintf("Collection %d", i+1),
-			Slug:         fmt.Sprintf("collection-%d", i+1),
-			DisplayOrder: int32(i),
-			Products:     &model.ProductConnection{Edges: []*model.ProductEdge{}, PageInfo: &model.PageInfo{}},
-			CreatedAt:    time.Now(),
-			UpdatedAt:    time.Now(),
-		}
-	}
-	return &model.ReorderCollectionsPayload{
-		ClientMutationID: input.ClientMutationID,
-		Collections:      collections,
-	}, nil
-}
 
 // PublishCatalog is the resolver for the publishCatalog field.
 func (r *mutationResolver) PublishCatalog(ctx context.Context, input model.PublishCatalogInput) (*model.PublishCatalogPayload, error) {
@@ -411,99 +59,6 @@ func (r *queryResolver) Nodes(ctx context.Context, ids []string) ([]model.Node, 
 		nodes[i] = node
 	}
 	return nodes, nil
-}
-
-// Product is the resolver for the product field.
-func (r *queryResolver) Product(ctx context.Context, by model.ProductBy) (*model.Product, error) {
-	if by.ID != nil {
-		productID, err := decodeNodeIDAs(nodeKindProduct, *by.ID)
-		if err != nil {
-			return nil, err
-		}
-		p, err := r.service.GetProductByID(ctx, productID)
-		if err != nil {
-			return nil, nil
-		}
-		return DatastoreProductToGraphQL(p), nil
-	}
-
-	p, err := r.service.GetProductBySKU(ctx, *by.Sku)
-	if err != nil {
-		return nil, nil
-	}
-	return DatastoreProductToGraphQL(p), nil
-}
-
-// Products is the resolver for the products field.
-func (r *queryResolver) Products(ctx context.Context, first *int32, after *string, last *int32, before *string) (*model.ProductConnection, error) {
-	params := toPageParams(first, after, last, before)
-	result, err := r.service.GetProducts(ctx, params)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get products: %w", err)
-	}
-	return BuildProductConnection(result), nil
-}
-
-// Category is the resolver for the category field.
-func (r *queryResolver) Category(ctx context.Context, by model.CategoryBy) (*model.Category, error) {
-	if by.ID != nil {
-		categoryID, err := decodeNodeIDAs(nodeKindCategory, *by.ID)
-		if err != nil {
-			return nil, err
-		}
-		c, err := r.service.GetCategoryByID(ctx, categoryID)
-		if err != nil {
-			return nil, nil
-		}
-		return DatastoreCategoryToGraphQL(c), nil
-	}
-
-	c, err := r.service.GetCategoryBySlug(ctx, *by.Slug)
-	if err != nil {
-		return nil, nil
-	}
-	return DatastoreCategoryToGraphQL(c), nil
-}
-
-// Categories returns categories as a Relay connection.
-func (r *queryResolver) Categories(ctx context.Context, first *int32, after *string, last *int32, before *string) (*model.CategoryConnection, error) {
-	params := toPageParams(first, after, last, before)
-	result, err := r.service.GetCategories(ctx, params)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get categories: %w", err)
-	}
-	return BuildCategoryConnection(result), nil
-}
-
-// Collection is the resolver for the collection field.
-func (r *queryResolver) Collection(ctx context.Context, by model.CollectionBy) (*model.Collection, error) {
-	if by.ID != nil {
-		collectionID, err := decodeNodeIDAs(nodeKindCollection, *by.ID)
-		if err != nil {
-			return nil, err
-		}
-		c, err := r.service.GetCollectionByID(ctx, collectionID)
-		if err != nil {
-			return nil, nil
-		}
-		return DatastoreCollectionToGraphQL(c), nil
-	}
-
-	c, err := r.service.GetCollectionBySlug(ctx, *by.Slug)
-	if err != nil {
-		return nil, nil
-	}
-	return DatastoreCollectionToGraphQL(c), nil
-}
-
-// Collections returns collections as a Relay connection.
-func (r *queryResolver) Collections(ctx context.Context, first *int32, after *string, last *int32, before *string) (*model.CollectionConnection, error) {
-	params := toPageParams(first, after, last, before)
-	result, err := r.service.GetCollections(ctx, params)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get collections: %w", err)
-	}
-	return BuildCollectionConnection(result), nil
 }
 
 // CatalogVersion is the resolver for the catalogVersion field.
