+++ gitstore-git-service/.env.example
@@ -10,7 +10,7 @@ GITSTORE_WS__PORT=8080                      # u16    — WebSocket notification 
 GITSTORE_GRPC__PORT=50051                   # u16    — gRPC service port
 GITSTORE_GIT__DATA_DIR=/data/repos          # string — path to repository storage
 GITSTORE_LOG__LEVEL=info                    # string — trace | debug | info | warn | error
-GITSTORE_LOG__FORMAT=json                   # string  — log format (json or text)
+GITSTORE_LOG__FORMAT=text                   # string  — log format (json or text)
 GITSTORE_GIT__REPO__MAX_FILE_SIZE=52428800  # u64    — max upload size in bytes (50 MiB)
 
 # ── Hook phase toggles (all disabled by default) ──────────────────────────────
